+++ convex/feedings.ts
@@ -1,10 +1,17 @@
-import { FeedingSchemaValidator } from '@feedingchart/app/feedingchart/model/feeding'
+import { FeedingSchemaValidator } from '@babylytics/app/src/model/feeding'
 import { mutation, query } from './_generated/server'
 import { v } from 'convex/values'
 
 export const get = query({
     args: { userId: v.string() },
     handler: async (ctx, args) => {
+        const user = await ctx.auth.getUserIdentity()
+        const userId = user?.tokenIdentifier.split('|')[1]
+
+        if (!user || userId !== args.userId) {
+            return []
+        }
+
         return (
             await ctx.db
                 .query('feedings')
@@ -17,13 +24,49 @@ export const get = query({
 export const getById = query({
     args: { id: v.optional(v.id('feedings')) },
     handler: async (ctx, args) => {
-        return args.id ? await ctx.db.get(args.id) : undefined
+        const user = await ctx.auth.getUserIdentity()
+        const userId = user?.tokenIdentifier.split('|')[1]
+
+        if (args.id == null) {
+            return undefined
+        }
+
+        const entry = await ctx.db.get(args.id)
+
+        if (!user) {
+            return Promise.reject('User not authenticated')
+        }
+
+        if (!entry) {
+            return Promise.reject('Feeding entry not found')
+        }
+
+        if (userId !== entry.userId) {
+            return Promise.reject(
+                `User token identifier ${userId} does not match feeding entry user ID ${entry.userId}`
+            )
+        }
+
+        return entry
     },
 })
 
 export const insert = mutation({
     args: { feeding: FeedingSchemaValidator },
     handler: async (ctx, args) => {
+        const user = await ctx.auth.getUserIdentity()
+        const userId = user?.tokenIdentifier.split('|')[1]
+
+        if (!user) {
+            return Promise.reject('User not authenticated')
+        }
+
+        if (userId !== args.feeding.userId) {
+            return Promise.reject(
+                `User token identifier ${userId} does not match diaper entry user ID ${args.feeding.userId}`
+            )
+        }
+
         return await ctx.db.insert('feedings', {
             ...args.feeding,
             time: args.feeding.time,
@@ -34,6 +77,30 @@ export const insert = mutation({
 export const update = mutation({
     args: { id: v.id('feedings'), feeding: FeedingSchemaValidator },
     handler: async (ctx, args) => {
+        const user = await ctx.auth.getUserIdentity()
+        const entry = await ctx.db.get(args.id)
+        const userId = user?.tokenIdentifier.split('|')[1]
+
+        if (!user) {
+            return Promise.reject('User not authenticated')
+        }
+
+        if (entry == null) {
+            return Promise.reject('Feeding entry not found')
+        }
+
+        if (entry.userId !== args.feeding.userId) {
+            return Promise.reject(
+                'Feeding entry user ID does not match provided feeding user ID'
+            )
+        }
+
+        if (userId !== args.feeding.userId) {
+            return Promise.reject(
+                `User token identifier ${userId} does not match diaper entry user ID ${args.feeding.userId}`
+            )
+        }
+
         return await ctx.db.patch<'feedings'>(args.id, args.feeding)
     },
 })
@@ -41,6 +108,24 @@ export const update = mutation({
 export const remove = mutation({
     args: { id: v.id('feedings') },
     handler: async (ctx, args) => {
+        const user = await ctx.auth.getUserIdentity()
+        const entry = await ctx.db.get(args.id)
+        const userId = user?.tokenIdentifier.split('|')[1]
+
+        if (!user) {
+            return Promise.reject('User not authenticated')
+        }
+
+        if (entry == null) {
+            return Promise.reject('Feeding entry not found')
+        }
+
+        if (userId !== entry.userId) {
+            return Promise.reject(
+                `User token identifier ${userId} does not match feeding entry user ID ${entry.userId}`
+            )
+        }
+
         return await ctx.db.delete(args.id)
     },
 })
